requirements.txt
streamlit==1.28.0
openai>=1.40.0
python-dotenv==1.0.0
pandas==2.3.1
# sqlite3
json5==0.9.14
markdown==3.5.1
# datetime
uuid